Add files via upload
Use this in place of the lines after line 99 in suma's previous upload to get the first 20 PC loadings and class labels all into a single data frame. Some R functions, such as 'neuralnet' require the data to be contained in a single df, so this may be useful for that :) 
Also, train and validate sets are created (with class labels attached as well) for later analysis.