Create LDA_QDA_KNN
Reads a CSV of 600 boat images resized for computational efficiency. Images have already been pre-processed to only include blue color intensities for each pixel. Images have been vectorized in the 'data setup' step such each image is represented by a 39,676 column vector corresponding to a resized/standardized image of size 115 by 345 pixels. 

For each image, PCA is conducted and the first 40 components. 40 was chosen in response to a comprehensive edge detection analysis that determined 90% pve can be obtained with this many components on these images. After PCA, an 80-20 training-to-validation set is sampled from the 600 image observations, and the image class labels are introduced. 

Five-fold cross validation is next used to select k for k-nearest neighbors. K-nearest-neighbors is executed and the test error rate is recorded. The accuracy (1 - misclassification rate) is decently accurate given the small sample size. 

Next, the assumption of normality is checked in preparation for potential LDA and QDA classification. Considering a histogram for each set of observations grouped by PC component, the PC scores appeared remarkably normal in all instances, and we thus concluded it appropriate to proceed with LDA and QDA.